consolidate packages and make NodeService ctor pkg private
