+++ core/src/main/java/org/elasticsearch/node/NodeModule.java
@@ -22,7 +22,6 @@ package org.elasticsearch.node;
 import org.elasticsearch.cluster.routing.allocation.DiskThresholdMonitor;
 import org.elasticsearch.common.inject.AbstractModule;
 import org.elasticsearch.monitor.MonitorService;
-import org.elasticsearch.node.service.NodeService;
 
 public class NodeModule extends AbstractModule {
